4_langgraph/community_contributions/langgraph-sidekick/pyproject.toml
[project]
name = "langgraph-sidekick"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "beautifulsoup4>=4.14.2",
    "gradio>=5.22.0",
    "langchain-community>=0.3.31",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.34",
    "langgraph>=0.3.18",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "lxml>=6.0.2",
    "playwright>=1.55.0",
    "python-dotenv>=1.1.1",
    "twilio>=9.8.4",
    "wikipedia>=1.4.0",
]
